+++ tbdex/rfq/rfq.go
@@ -1,4 +1,4 @@
-package tbdex
+package rfq
 
 import (
 	"crypto/sha256"
